Move SplatDate to its own file
